Fused virtual machine: add the possibility to highlight machines
Previously all the vm where highlighted and the areas corresponding to the
host had their colors changed to a dull color.
Now with the M button as in Machines added in the window we can select
which machine we want to highlight or not.

The next step is to add the possibility to choose a specific cpu or vcpu
instead of always all the machine.