fused vm: adds new way to select what to highlight
It is now possible to highlight a machine (real or virtual), a cpu, or a
process. You just need to press the corresponding button in the tool bar.
The effect of those buttons depends on where you clicked in the view.
A tooltip text on each button reminds what is currently selected.

It is also possible to select a process to highlight by clicking on it in
the Control FLow View. Even if it is a process in a virtual machine.