fused vm: corrects a bug in the machine's cpus detection
If a trace of a vm was longer than the host's trace a pCPU could be
detected as a vCPU and could appear in the select machine dialog. This has
been corrected.

All the code related to the select machine dialog is now only inside the
plugin. Exept for the image of the button but it's seems to do not work
right now.