fused vm: add optimisations for the view
Highlighting things is no longer done by duplicating states but by
overriding getEventAlpha.
Anybody making a timegraph view can highlight or not some states.

A hashmap is used to remember the result of a highlight query on the state
system depending on a timeevent. Displaying the rows is now faster.

The color of a rectangle using an alpha coefficient is computed before
drawing the rectangle by assuming that the previous color was white.