fused vm: allow shades of colors by scrolling
The view can now modify the shade of the color used for a specific process.
By pushing shift and scrolling after selecting a process in the view you
can modify the shade of the color.
This allows to see more than two different highlighted threads at the same
time.

The analysis is less dependent of the kernel analysis.
This solves a problem of restoring states after an IRQ and a bug where the view
didn't know cpu was back in kernel mode after being in guest mode and in a
syscall.