fused vm: Add support of containers in the view
It is possible to highlight specific containers.
The selection is done with the Select Machine button. The list of machines will
contain all the detected namespaces. Including the root namespace.
Currently this version doesn't see the difference between two namespaces with
the same ID. This can happen with containers on multiple machines.

The algorithm to handle the statedump_process events was rewritten. It creates
empty attributes in the sht that can be filled later.