fused vm: removed the need of transition states
Because the view was creating time events only when there was a change of cpu
status, the filter couldn't directly work and some false satus were used to
express a change of thread or machine on one cpu.
This has been corected in the view so the analysis does not depend on that
anymore.

This commit fixes also a bug where things happening on a vcpuX where merged
to pcpuX when the real pcpu was still unknown.
Now the events are just ignored in this case.