fused vm: add support of nested vms in qemukvm model
A machine can now be a host and a guest at the same time.
The analysis doesn't really start until all roles for the machines are found.
With the exception of events required from the lxc model.

Each machine knows its parent, it's like for the containers.
Consequently the view can recreate the hierarchy and display it.

Attention, the exits and entries from L2 are not implemented yet.
So currently the view doesn't display anything relevant.