fused vm: support of entries and exits to L2
Implementation of the algorithm to detect when a pcpu is running a
vcpu from L2.
It gives good results in the view if L2 is correctly synchronized with L0.

In the view the pcpu entries of a vm show also the activity of the vm contained
inside this vm. Consequently vm's entries are a subset of the machine
containing it.